Failable `tryOver` for `Prism`s and `Traversals`
a.k.a. `(%%~?)`

The default behavior for over applied to Prisms, to act as id if the
prismatic mapping fails, is concerning since it consumes a failure.
Originally a failing `over` combinator was defined in terms of `over`
and `isn't`, but a more general combinator can be defined over
`Traversal`s using `%%~` and `Any`.