Make Equality more poly-kinded
Allow `Equality s t a b` to prove `s ~ a` and `t ~ b` whenever
`s` and `a` have the same kind and `t` and `b` have the same kind.

Fixes #605

Note: For GHC 7.8 and later, this changes the type of `AnEquality`
in a non-backwards-compatible way, but this likely won't affect
much code.