Added indexed versions of minimumOf etc.
These functions are the analogue of `minimumOf`, `minimumByOf`,
`maximumOf` and `maximumByOf` which also return the position of the
element in the indexed fold.

This could be useful, for example, when trying to modify the minimum
element in a fold, when one doesn't want to expose a bidirectional
optics for the bad compositional properties, eg `modifyMin f . modifyMin
g /= modifyMin (f . g)`.