Add type families for concrete optics
It's occasionally useful to get one's hands on the concrete `Functor` or
`Profunctor` of an `ALens`, `APrism`, etc., to use in constraints. Currently,
the only built-in option is to import the relevant `.Internal` module and use
the right one. But since these types are all concrete and transparent, it's
actually possible to write type families to extract the relevant types.
So let's do that, purely for convenience.

Note: I believe it must be possible to use `TypeInType` to make the kinds
of the results depend on the arguments. So we should be able to ask for
something like `GetConcreteFunctor (AnEquality 'Bool 'Int 'Bool 'Int)`.
But I don't yet know enough about `TypeInType` to be able to do that myself.