Move FieldNamer, ClassyNamer to Control.Lens.Internal.FieldTH
It took me a while to realize that two of the fields in LensRules were actually
of types FieldNamer and ClassyNamer, since these type synonyms were defined in
a separate module from LensRules. This moves FieldNamer and ClassyNamer to the
same module as LensRules and uses these types in the record fields of LensRules.