Disambiguate TH-generated classy prisms for punned data type names
Consider a scenario in which one uses Template Haskell to generate
classy prisms for a data type that shares a name with one of its
constructors, as in the following example from #865:

```hs
newtype SDLException = SDLException String deriving Show
makeClassyPrisms ''SDLException
```

We clearly can't use the name `_SDLException` to refer to both the
prism for the data type and the prism for the constructor, so we need
to disambiguate them somehow. This patch adopts the convention that
in such a scenario, the name of the prism for the data type will be
prefixed with an extra `_` (i.e., `__SDLException`) to distinguish it
from the constructor prism, `_SDLException`.

Fixes #865.