Add more data type prisms to Control.Exception.Lens (#871)
This is a follow-up to #869 that implements the proposed refactoring
to the classy prisms in `Control.Exception.Lens` stated in
https://github.com/ekmett/lens/issues/865#issuecomment-501368037.
For each class dedicated to a data type that shares a name with one
of its constructors (e.g., `AsNonTermination`), we add a data type
prism prefixed with `__`, following the convention established in
867d41f76043b6909d13959d415d24bc2af64952. We also define default
implementations for the existing prisms in terms of the new data
type prism.

In the strictest sense, this is a backwards-compatible change, as
existing code that defines instances of `AsNonTermination` _et al._
will continue to compile. However, this code will now produce
warnings stating that `__AsNonTermination` is not defined, so this
will likely inspire people to update their code anyway. Luckily, this
change will debut in a new major version of `lens`, so it's easier
to justify.