Re-enable -Wunused-matches (#899)
Commit b2e0efe6842f25c545291579baecf8790ee24636 added
`-Wno-unused-matches` to `lens.cabal` in an effort to work around
GHC#11451, which affected GHC 8.0.1-rc1 (the newest compiler at the
time). However, #11451 has since been fixed, and as far as I can
tell, it does not affect the latest point release of GHC 8.0
(8.0.2). In light of this, it seems a bit silly to me to specifically
disable this warning when it no longer provides any tangible benefit.
Let's reenable it.