Remove the test suite's dependency on test-framework-th
This is primarily motivated by `test-framework-th`'s dependency on
`language-haskell-extract`, which does not build on GHC 8.10+ and
appears to be abandoned by its maintainer. See the discussion
at #918 for more details.

Removing this dependency finally allows us to build `lens` with GHC
8.10.1 on Travis without the need for `head.hackage`, so I took the
liberty of regenerating `.travis.yml` in this patch as well.

Fixes #918.