Use strict and assoc
Fixes #936. We also depend on assoc for Swap(ped), as `strict`
depends on `assoc` to provide `Swap` instances.

I hope that some day `Swap` class will migrate to `bifunctors`
or even to `base`.

There is `strict-lens`, with `Data.Strict.Lens` module,
adding some optics for types in `Data.Strict`.
It can be merged into `lens` or kept separate.

I also added a dependency to `these` to provide an Each instance
these is a dependency of strict.
(The dependency graph could be refined, but not having orphans
is a good start)