Enable -Wmissing-pattern-synonym-signatures
Now that we require GHC 8.0 as the minimum, I no longer see any reason to
use `-Wno-missing-pattern-synonym-signatures`, as we can always define pattern
synonym signatures without CPP. In fact, we already do so. We might as well
enable the corresponding warning to make sure this remains the case in the
future.