Allow {make,declare}Wrapped to work with data family instances
This modifies the implementations of `makeWrapped` and `declareWrapped` to
use `th-abstraction`, which grants the ability to work with both ordinary data
types and data family instances. An added bonus of doing this is that we can
delete a fair bit of delicate, CPP-laden code that only existed to support the
old implementations of `makeWrapped` and `declareWrapped`.

Fixes #997.