Disambiguate Haddock references
Due to a bug, references to e.g. 'Lens' in the Haddock documentation are
resolved to ReifiedLens if the Lens value constructor is in scope.

Fully annotating as 'Control.Lens.Lens.Lens' solves the issue visually
but generates broken links (another bug).

Haddock 2.23 allows us to explicitly namespace those type constructors
as types, which solves the issue completely.