Avoid -Wx-partial warnings in Data.Data.Lens's doctests
Some doctests in `Data.Data.Lens` use the partial `head` and `tail` functions,
which now produce `-Wx-partial` warnings (implied by `-Wall`) as of GHC 9.8.
Moreover, some of these doctests _rely_ on `head`/`tail` being partial, so it's
not entirely straightforward to change them. To accommodate this, we redefine
the `head`/`tail` functions (see the new `Control.Lens.Internal.Doctest`
module) so that they can be used in the doctests without triggering any
warnings. The end result should be indistinguishable to readers of the
Haddocks.