feat(Setter): add: prepends setter
Add a setter that prepends the element, like `cons`.
For balance, we also add a setter that also appends the element to the tail, like `snoc`.
We often use this when adding elements to list structures and other order-conscious things.
When we are doing natural language processing, for example, and we are adding processed data, it is more natural to add the resulting processed data to the head side if the original input source is closer to the head.
Naturally we can also construct the processing by adding the tail if we think hard enough, but we don't want to work hard at that since we only need to prepare a function.
The naming of the `<>:~` function was lost with `~<>`, but considering that the `~` symbol in the lenses setter operator is basically on the right, I used the `:` operator to indicate that it is closer to the `cons` of the list, rather than focusing on the contrast.