Benchmarks: Remove base-compat dependency
The `plated` benchmark suite previously had a dependency on `base-compat`,
which was introduced in #909 for compatibility with pre-7.10 versions of GHC.
In #977, however, the minimum version of GHC was raised to 8.0, which means
that there is no longer any reason for these benchmarks to depend on
`base-compat`. This patch removes the dependency.