Reduce arity of foldlOf' and related functions
This is done in order to make GHC inliner more eager to inline them and do that
earlier. Standard heuristic is to inline function when all arguments are applied.
For example this means foldlOf' will be inlined into body of sumOf and this
gives GHC more opportunities for optimization

Simple benchmark for summong list using sumOf see 8x performance improvement

Fixed #1084

Co-authored-by: Ryan Scott <ryan.gl.scott@gmail.com>